+++ WorldInterface.sol
@@ -6,21 +6,21 @@ interface WorldInterface {
      * @param lieu le numéro du lieu
      * @return vrai si le lieu existe, faux si ce n'est pas le cas
      */
-    function estLieuValide(uint16 lieu) external pure returns(bool);
+    function estLieuValide(uint16 lieu) external pure returns (bool);
 
     /**
      * Donne le coût du lieu en ether.
      * @param lieu le numéro du lieu
      * @return le coût du lieu en ether
      */
-    function getCoutLieu(uint16 lieu) external view returns(uint);
+    function getCoutLieu(uint16 lieu) external view returns (uint256);
 
     /**
      * Vérifie si le lieu est occupé.
      * @param lieu le numéro du lieu
      * @return vrai si le lieu est occupé, faux s'il est libre.
      */
-    function estOccupe(uint16 lieu) external view returns(bool);
+    function estOccupe(uint16 lieu) external view returns (bool);
 
     /**
      * Explore un lieu avec un avatar. (coûte des ethers suivant le lieu)
@@ -29,7 +29,11 @@ interface WorldInterface {
      * @param codeAcces le code d'accès au lieu, si le code est incorrect l'exploration echouera levant une erreur.
      * @return un code d'accès à d'autres lieux
      */
-    function explorer(uint16 lieu, address avatar, uint64 codeAcces) external payable returns (uint64);
+    function explorer(
+        uint16 lieu,
+        address avatar,
+        uint64 codeAcces
+    ) external payable returns (uint64);
 
     /**
      * Fait quitter le lieu à un avatar.
@@ -51,5 +55,9 @@ interface WorldInterface {
      * @param adrEscouade l'adresse du smarcontract de l'escouade de l'avatar présent sur le lieu
      * @param adrAvatar l'adresse du smarcontrat de l'avatar présent sur le lieu
      */
-    function appliquerEffetLieu(uint16 lieu, address adrEscouade, address adrAvatar) external;
+    function appliquerEffetLieu(
+        uint16 lieu,
+        address adrEscouade,
+        address adrAvatar
+    ) external;
 }
